move remote state
