+++ services/requester/remote_state.go
@@ -1,4 +1,4 @@
-package emulator
+package requester
 
 import (
 	"context"
@@ -15,7 +15,7 @@ import (
 	"google.golang.org/grpc/status"
 )
 
-var previewnetStorage = flow.HexToAddress("0x4f6fd534ddd3fc5f")
+var previewnetStorageAddress = flow.HexToAddress("0x4f6fd534ddd3fc5f")
 
 var _ atree.Ledger = &remoteLedger{}
